Tweak Date data types; Fix type errors; Refactor
