--- ebl/fragmentarium/application/date_range_schema.py
@@ -1,7 +1,0 @@
-from marshmallow import Schema, fields
-
-
-class DateRangeSchema(Schema):
-    start = fields.Date()
-    end = fields.Date()
-    notes = fields.String()
